requirements.txt
numpy
tensorflow
Pillow
keras
flask